refactor(lowy): move agentKindFromCommand into anyagent
The basename→AgentKind bridge lives where the basename axis is already
owned: `STABLE_FLAGS` and `parseAgentCommand` in
`anyagent/src/agent-cli.ts`. Previously the client carried a shadow
table (`COMMAND_BASENAME_TO_KIND`) that had to stay in sync with
anyagent's enumeration — adding a fourth icon-capable agent meant
editing two packages where one would do.

Anyagent exports a literal-union `AgentKind` matching `AgentInfo["kind"]`
in kolu-common; kolu-common already depends on anyagent so the type
direction is consistent. The client's `agentDisplay.ts` returns to a
pure display module (icons + names + state labels). Tests for
`agentKindFromCommand` move next to `parseAgentCommand`'s.
